move tests regarding implied bounds to their own file
